M- design + bugfix

+++ src/main/java/glp/controller/ModerationController.java
@@ -44,79 +44,83 @@ public class ModerationController {
 	
 	@RequestMapping("/valider/annonce/{id}")
 	public ModelAndView validerAnnonce(@PathVariable("id") int idAnnSelected) {
-//		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
 			Annonce ann = annonceService.getRowById(idAnnSelected);
 			ann.setValide(true);
 			annonceService.updateRow(ann);
-//		}
-		return getListAnnoncesAModerer();
+			return getListAnnoncesAModerer();
+		}  else
+			return new ModelAndView("redirect:/");
 	}
 	
 	@RequestMapping("/refuser/annonce/{id}")
 	public ModelAndView refuserAnnonce(@PathVariable("id") int idAnnSelected,
 			HttpServletRequest request) {
-//		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
 			Annonce ann = annonceService.getRowById(idAnnSelected);
 			ann.setValide(false);
-			
 			String motif = request.getParameter("motif");
 			String content = "Bonjour, votre annonce intitulée : " + ann.getTitre() + "\n\n"
 					 + "dont la description est la suivante \n\n" + ann.getDesc() + "\n\n\n"
 					 + "a été modérée pour le motif suivant : " + motif;
-			
-			
-			
 			EmailSender.sendMail(ann.getAuteur().getMailLille1(), "Lille1Community", content, ann);
-			
-			
 			annonceService.deleteRow(ann.getId());
-//		}
-		return getListAnnoncesAModerer();
+			return getListAnnoncesAModerer();
+		}  else
+			return new ModelAndView("redirect:/");
 	}
 	
 	@RequestMapping("list")
 	public ModelAndView getListAnnoncesAModerer() {
-		Map<String, Object> modelAnnoncesAModerer = new HashMap<String, Object>();
-		modelAnnoncesAModerer.put("utilisateur", utilisateurService.getUserInSession());
-		modelAnnoncesAModerer.put("catList", categorieService.getList());
-//		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+			Map<String, Object> modelAnnoncesAModerer = new HashMap<String, Object>();
+			modelAnnoncesAModerer.put("utilisateur", utilisateurService.getUserInSession());
+			modelAnnoncesAModerer.put("catList", categorieService.getList());
 			modelAnnoncesAModerer.put("annonceList", annonceService.getListAModerer());
-//		}
-		return new ModelAndView("mod_listeAnnonces", modelAnnoncesAModerer);
+			return new ModelAndView("mod_listeAnnonces", modelAnnoncesAModerer);
+		}  else
+			return new ModelAndView("redirect:/");
 	}
 	
 	@RequestMapping(value ="", method = RequestMethod.GET)
 	public ModelAndView afficheStatistique(){
-		Map<String, Object> myModel = new HashMap<String, Object>();
-//		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+			Map<String, Object> myModel = new HashMap<String, Object>();
 			int nbUser = utilisateurService.nbUtilisateur();
 			int nbAnnonceEnligne= annonceService.nbAnnonceEnLigne();		
 			Stats stats = annonceService.getStats();
 			Map<String, Integer> annonceByCat = annonceService.getNbByCategorie();
+			myModel.put("dureeVieAnnonce", stats.getNb_jours_fin_annonce());
+			myModel.put("utilisateur", utilisateurService.getUserInSession());
 			myModel.put("annonceByCat", annonceByCat);
 			myModel.put("nbUser", nbUser);
 			myModel.put("nbAnnonceEnligne", nbAnnonceEnligne);
 			myModel.put("nbAnnCrees", stats.getStats_nb_ann_crees());
 			myModel.put("stats", stats);
-// 		}
-		return new ModelAndView("statistique", myModel);
+			return new ModelAndView("statistique", myModel);
+		}  else
+			return new ModelAndView("redirect:/");
 	}
 	
 	@RequestMapping(value="setDureeVieAnn", method=RequestMethod.GET)
 	public ModelAndView setDureeVieAnnonce(@RequestParam("duree_vie") int duree_vie_ann) {
-		annonceService.setDureeVieAnnonce(duree_vie_ann);
-		return afficheStatistique();
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+			annonceService.setDureeVieAnnonce(duree_vie_ann);
+			return afficheStatistique();
+		}  else
+			return new ModelAndView("redirect:/");
 	}
 	
 	@RequestMapping("/signalement/list")
 	public ModelAndView getListSignalements() {
-		Map<String, Object> modelAnnoncesAModerer = new HashMap<String, Object>();
-		modelAnnoncesAModerer.put("utilisateur", utilisateurService.getUserInSession());
-		modelAnnoncesAModerer.put("catList", categorieService.getList());
-//		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+		if(utilisateurService.isModerateur(utilisateurService.getUserInSession())) {
+			Map<String, Object> modelAnnoncesAModerer = new HashMap<String, Object>();
+			modelAnnoncesAModerer.put("utilisateur", utilisateurService.getUserInSession());
+			modelAnnoncesAModerer.put("catList", categorieService.getList());
 			modelAnnoncesAModerer.put("signalements", signalisationService.getList());
-//		}
-		return new ModelAndView("mod_listeSignalements", modelAnnoncesAModerer);
+			return new ModelAndView("mod_listeSignalements", modelAnnoncesAModerer);
+		}  else
+			return new ModelAndView("redirect:/");
 	}
